+++ spark-bigquery-dsv1/src/test/scala/com/google/cloud/spark/bigquery/SchemaIteratorSuite.scala
@@ -18,6 +18,10 @@ package com.google.cloud.spark.bigquery
 import com.google.cloud.bigquery.Field.Mode
 import com.google.cloud.bigquery.LegacySQLTypeName.{BOOLEAN, BYTES, DATE, DATETIME, FLOAT, INTEGER, JSON, NUMERIC, RECORD, STRING, TIME, TIMESTAMP}
 import com.google.cloud.bigquery.{Field, Schema}
+import com.google.cloud.bigquery.storage.v1.AvroRows;
+import com.google.cloud.bigquery.storage.v1.ReadRowsResponse;
+import com.google.cloud.bigquery.storage.v1.ReadSession.TableReadOptions.ResponseCompressionCodec;
+import com.google.cloud.bigquery.storage.v1.ArrowRecordBatch;
 import com.google.common.io.ByteStreams.toByteArray
 import com.google.protobuf.ByteString
 import org.apache.avro.{Schema => AvroSchema}
@@ -30,6 +34,8 @@ import org.scalatest.funsuite.AnyFunSuite
 import java.util.Optional
 import scala.collection.JavaConverters._
 
+
+
 /**
  * A test for ensuring that Arrow and Avros Schema generate same results for
  * underlying BigQuery source
@@ -39,6 +45,7 @@ class SchemaIteratorSuite extends AnyFunSuite {
 
   val schemaConvertersConfiguration = SchemaConvertersConfiguration.createDefault()
   val schemaConverters = SchemaConverters.from(schemaConvertersConfiguration)
+  val responseCompressionCodec = ResponseCompressionCodec.RESPONSE_COMPRESSION_CODEC_UNSPECIFIED
 
   test("compare arrow and avro results") {
     // rows in the form of bytes string in both arrow and avro format
@@ -46,6 +53,13 @@ class SchemaIteratorSuite extends AnyFunSuite {
       toByteArray(getClass.getResourceAsStream("/alltypes.avro")))
     val arrowByteString = ByteString.copyFrom(
       toByteArray(getClass.getResourceAsStream("/alltypes.arrow")))
+
+    // construct a ReadRowsResponse to pass to AvroBinaryIterator
+    val avroReadRowsResponse =  ReadRowsResponse.newBuilder().setAvroRows(
+      AvroRows.newBuilder().setSerializedBinaryRows(avroByteString)).build();
+    val arrowReadRowsResponse = ReadRowsResponse.newBuilder().setArrowRecordBatch(
+      ArrowRecordBatch.newBuilder().setSerializedRecordBatch(arrowByteString)
+    ).build();
 
     // avro and arrow schemas required to read rows from bigquery
     val arrowSchema = ByteString.copyFrom(toByteArray(getClass.getResourceAsStream("/alltypes.arrowschema")))
@@ -82,9 +96,10 @@ class SchemaIteratorSuite extends AnyFunSuite {
 
     val arrowBinaryIterator = new ArrowBinaryIterator(columnsInOrder.asJava,
       arrowSchema,
-      arrowByteString,
+      arrowReadRowsResponse,
       Optional.empty(),
-      null).asScala
+      null,
+      responseCompressionCodec).asScala
 
     if (arrowBinaryIterator.hasNext) {
        arrowSparkRow = arrowBinaryIterator.next();
@@ -93,10 +108,11 @@ class SchemaIteratorSuite extends AnyFunSuite {
     val avroBinaryIterator = new AvroBinaryIterator(bqSchema,
       columnsInOrder.asJava,
       avroSchema,
-      avroByteString,
+      avroReadRowsResponse,
       Optional.empty(),
       Optional.empty(),
-      schemaConvertersConfiguration)
+      schemaConvertersConfiguration,
+      responseCompressionCodec)
 
     if (avroBinaryIterator.hasNext) {
       avroSparkRow = avroBinaryIterator.next()
@@ -156,6 +172,13 @@ class SchemaIteratorSuite extends AnyFunSuite {
     val arrowByteString = ByteString.copyFrom(
       toByteArray(getClass.getResourceAsStream("/alltypes.arrow")))
 
+    // construct a ReadRowsResponse to pass to AvroBinaryIterator
+    val avroReadRowsResponse =  ReadRowsResponse.newBuilder().setAvroRows(
+      AvroRows.newBuilder().setSerializedBinaryRows(avroByteString)).build();
+    val arrowReadRowsResponse = ReadRowsResponse.newBuilder().setArrowRecordBatch(
+      ArrowRecordBatch.newBuilder().setSerializedRecordBatch(arrowByteString)
+    ).build();
+
     // avro and arrow schemas required to read rows from bigquery
     val arrowSchema = ByteString.copyFrom(toByteArray(getClass.getResourceAsStream("/alltypes.arrowschema")))
     val avroSchema = new AvroSchema.Parser().
@@ -191,9 +214,10 @@ class SchemaIteratorSuite extends AnyFunSuite {
 
     val arrowBinaryIterator = new ArrowBinaryIterator(columnsInOrder.asJava,
       arrowSchema,
-      arrowByteString,
+      arrowReadRowsResponse,
       Optional.empty(),
-      Optional.empty()).asScala
+      Optional.empty(),
+      responseCompressionCodec).asScala
 
     if (arrowBinaryIterator.hasNext) {
       arrowSparkRow = arrowBinaryIterator.next();
@@ -202,10 +226,11 @@ class SchemaIteratorSuite extends AnyFunSuite {
     val avroBinaryIterator = new AvroBinaryIterator(bqSchema,
       columnsInOrder.asJava,
       avroSchema,
-      avroByteString,
+      avroReadRowsResponse,
       Optional.empty(),
       Optional.empty(),
-      schemaConvertersConfiguration)
+      schemaConvertersConfiguration,
+      responseCompressionCodec)
 
     if (avroBinaryIterator.hasNext) {
       avroSparkRow = avroBinaryIterator.next()
@@ -261,6 +286,9 @@ class SchemaIteratorSuite extends AnyFunSuite {
     val arrowByteString =
       ByteString.copyFrom(
         toByteArray(getClass.getResourceAsStream("/arrowDateTimeRowsInBytes")))
+    val arrowReadRowsResponse = ReadRowsResponse.newBuilder().setArrowRecordBatch(
+      ArrowRecordBatch.newBuilder().setSerializedRecordBatch(arrowByteString)
+    ).build();
 
     val arrowSchema =
       ByteString.copyFrom(
@@ -270,7 +298,12 @@ class SchemaIteratorSuite extends AnyFunSuite {
 
     val arrowBinaryIterator =
       new ArrowBinaryIterator(
-        columnsInOrder.asJava, arrowSchema, arrowByteString, Optional.empty(), Optional.empty()).asScala
+        columnsInOrder.asJava,
+        arrowSchema,
+        arrowReadRowsResponse,
+        Optional.empty(),
+        Optional.empty(),
+        responseCompressionCodec).asScala
 
     while (arrowBinaryIterator.hasNext) {
       val arrowSparkRow = arrowBinaryIterator.next()